requirements.txt
fastapi==0.116.1
fastapi_mcp==0.4.0
langchain_core==0.3.74
langgraph==0.6.5
langsmith==0.4.14
openai==1.99.9
pydantic==2.11.7
python-dotenv==1.1.1
toml==0.10.2
uvicorn==0.35.0
